refactor: moving header components into subpackage, moving cmd opts and flags to node/header
