+++ node/tests/sync_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/celestiaorg/celestia-node/node"
+	headerconf "github.com/celestiaorg/celestia-node/node/header"
 	nodeconf "github.com/celestiaorg/celestia-node/node/node"
 	"github.com/celestiaorg/celestia-node/node/tests/swamp"
 )
@@ -49,7 +50,7 @@ func TestSyncLightWithBridge(t *testing.T) {
 	addrs, err := peer.AddrInfoToP2pAddrs(host.InfoFromHost(bridge.Host))
 	require.NoError(t, err)
 
-	light := sw.NewLightNode(nodeconf.WithTrustedPeers(addrs[0].String()))
+	light := sw.NewLightNode(headerconf.WithTrustedPeers(addrs[0].String()))
 
 	err = light.Start(ctx)
 	require.NoError(t, err)
@@ -98,7 +99,7 @@ func TestSyncStartStopLightWithBridge(t *testing.T) {
 	require.NoError(t, err)
 
 	store := node.MockStore(t, nodeconf.DefaultConfig(nodeconf.Light))
-	light := sw.NewNodeWithStore(nodeconf.Light, store, nodeconf.WithTrustedPeers(addrs[0].String()))
+	light := sw.NewNodeWithStore(nodeconf.Light, store, headerconf.WithTrustedPeers(addrs[0].String()))
 	require.NoError(t, light.Start(ctx))
 
 	h, err = light.HeaderServ.GetByHeight(ctx, 30)
@@ -109,7 +110,7 @@ func TestSyncStartStopLightWithBridge(t *testing.T) {
 	require.NoError(t, light.Stop(ctx))
 	require.NoError(t, sw.RemoveNode(light, nodeconf.Light))
 
-	light = sw.NewNodeWithStore(nodeconf.Light, store, nodeconf.WithTrustedPeers(addrs[0].String()))
+	light = sw.NewNodeWithStore(nodeconf.Light, store, headerconf.WithTrustedPeers(addrs[0].String()))
 	require.NoError(t, light.Start(ctx))
 
 	h, err = light.HeaderServ.GetByHeight(ctx, 40)
@@ -149,7 +150,7 @@ func TestSyncFullWithBridge(t *testing.T) {
 	addrs, err := peer.AddrInfoToP2pAddrs(host.InfoFromHost(bridge.Host))
 	require.NoError(t, err)
 
-	full := sw.NewFullNode(nodeconf.WithTrustedPeers(addrs[0].String()))
+	full := sw.NewFullNode(headerconf.WithTrustedPeers(addrs[0].String()))
 	require.NoError(t, full.Start(ctx))
 
 	h, err = full.HeaderServ.GetByHeight(ctx, 30)
@@ -195,7 +196,7 @@ func TestSyncLightWithFull(t *testing.T) {
 	addrs, err := peer.AddrInfoToP2pAddrs(host.InfoFromHost(bridge.Host))
 	require.NoError(t, err)
 
-	full := sw.NewFullNode(nodeconf.WithTrustedPeers(addrs[0].String()))
+	full := sw.NewFullNode(headerconf.WithTrustedPeers(addrs[0].String()))
 	require.NoError(t, full.Start(ctx))
 
 	h, err = full.HeaderServ.GetByHeight(ctx, 30)
@@ -206,7 +207,7 @@ func TestSyncLightWithFull(t *testing.T) {
 	addrs, err = peer.AddrInfoToP2pAddrs(host.InfoFromHost(full.Host))
 	require.NoError(t, err)
 
-	light := sw.NewLightNode(nodeconf.WithTrustedPeers(addrs[0].String()))
+	light := sw.NewLightNode(headerconf.WithTrustedPeers(addrs[0].String()))
 
 	err = sw.Network.UnlinkPeers(bridge.Host.ID(), light.Host.ID())
 	require.NoError(t, err)
@@ -259,7 +260,7 @@ func TestSyncLightWithTrustedPeers(t *testing.T) {
 
 	trustedPeers := []string{addrs[0].String()}
 
-	full := sw.NewFullNode(nodeconf.WithTrustedPeers(addrs[0].String()))
+	full := sw.NewFullNode(headerconf.WithTrustedPeers(addrs[0].String()))
 	require.NoError(t, full.Start(ctx))
 
 	h, err = full.HeaderServ.GetByHeight(ctx, 30)
@@ -272,7 +273,7 @@ func TestSyncLightWithTrustedPeers(t *testing.T) {
 
 	trustedPeers = append(trustedPeers, addrs[0].String())
 
-	light := sw.NewLightNode(nodeconf.WithTrustedPeers(trustedPeers...))
+	light := sw.NewLightNode(headerconf.WithTrustedPeers(trustedPeers...))
 
 	err = light.Start(ctx)
 	require.NoError(t, err)
